Modules/downloader/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

httpx==0.27.2
structlog==24.4.0

ffmpeg-python==0.2.0

# Этап 4: живые стратегии
yt-dlp==2026.3.17
# Apify вызываем через httpx (лёгкий + матчит паттерн monitor), без apify-client SDK.
